Fix some compile warnings

+++ examples/common/tengine_operations.c
@@ -589,7 +589,7 @@ image copy_image(image p)
 {
     image copy = p;
     copy.data = ( float* )calloc((size_t)p.h * p.w * p.c, sizeof(float));
-    memcpy(copy.data, p.data, p.h * p.w * p.c * sizeof(float));
+    memcpy(copy.data, p.data, (unsigned long)p.h * p.w * p.c * sizeof(float));
     return copy;
 }
